requirements.txt
flask
line-bot-sdk
apscheduler
python-dotenv
requests
